Refactor users
