+++ src/reservations/new-reservation/reservationRequest.validator.ts
@@ -1,5 +1,5 @@
 import { getValidationMsg } from "@src/controllerUtilities"
-import { NewReservationDto } from "@src/reservations/interface/ReservationInterfaceDto"
+import { NewReservationDto } from "@src/reservations/interface/reservation.interface.dto"
 import dayjs from "dayjs"
 import { NextFunction, Request, Response } from "express"
 import { StatusCodes } from "http-status-codes"
@@ -29,4 +29,3 @@ export const validateNewReservation = (request: Request, res: Response, next: Ne
   }
   next();
 }
-
