+++ src/users/user-login/login.service.ts
@@ -1,0 +1,8 @@
+import { User, UserCrendentials } from "../interface/user.interface"
+
+export const loginService = {
+
+  login: (userCredentials: UserCrendentials): User | undefined => {
+    throw new Error("Not implemented")
+  }
+}
