+++ src/users/user.mapper.ts
@@ -1,5 +1,5 @@
-import { User } from "@src/users/interface/UserInterface";
-import { UserDto } from "@src/users/interface/UserInterfaceDto";
+import { User } from "@src/users/interface/user.interface";
+import { UserDto } from "@src/users/interface/user.interface.dto";
 
 export const userMapper = {
   toDto: (user: User): UserDto => {
